Added better error handling for like controller

+++ IzvorniKod/backend/src/main/java/com/PROGI/backend/dao/LikesDataAccessService.java
@@ -1,31 +1,36 @@
 package com.PROGI.backend.dao;
 
+import com.PROGI.backend.exceptions.RecipeNotFound;
+import com.PROGI.backend.exceptions.ProfileNotFound;
 import com.PROGI.backend.mappers.LikeMapper;
 import com.PROGI.backend.mappers.RecipeMapper;
 import com.PROGI.backend.model.Like;
 import com.PROGI.backend.model.Recipe;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.List;
 import java.util.UUID;
 
 @Repository("postgresLike")
 public class LikesDataAccessService implements LikesDao {
     private final JdbcTemplate jdbcTemplate;
+    private final RecipeDao recipeDao;
+    private final ProfileDao profileDao;
 
     @Autowired
-    public LikesDataAccessService(JdbcTemplate jdbcTemplate) {
+    public LikesDataAccessService(JdbcTemplate jdbcTemplate, @Qualifier("postgresRecipe") RecipeDao recipeDao, @Qualifier("postgresProfile") ProfileDao profileDao) {
         this.jdbcTemplate = jdbcTemplate;
+        this.recipeDao = recipeDao;
+        this.profileDao = profileDao;
     }
 
     @Override
-    public void addLike(Like like) {
+    public void addLike(Like like) throws RecipeNotFound, ProfileNotFound {
+        if(recipeDao.selectRecipeById(like.getRecipeId()).isEmpty()) throw new RecipeNotFound(like.getRecipeId());
+        if(profileDao.selectProfileById(like.getUserId()).isEmpty()) throw new ProfileNotFound(like.getUserId());
         String sql = "INSERT INTO likes(userId, recipeId) VALUES(?, ?)";
         jdbcTemplate.update(sql, like.getUserId(), like.getRecipeId());
     }
@@ -37,13 +42,16 @@ public class LikesDataAccessService implements LikesDao {
     }
 
     @Override
-    public List<Recipe> getLikedRecipes(UUID uid) {
+    public List<Recipe> getLikedRecipes(UUID uid) throws ProfileNotFound {
+        if(profileDao.selectProfileById(uid).isEmpty()) throw new ProfileNotFound(uid);
         String sql = "SELECT * FROM likes l JOIN recipe r ON l.recipeId = r.recipeId WHERE l.userid = ?";
         return jdbcTemplate.query(sql, new RecipeMapper(), uid.toString());
     }
 
     @Override
-    public void deleteLike(UUID uid, UUID rid) {
+    public void deleteLike(UUID uid, UUID rid) throws RecipeNotFound, ProfileNotFound {
+        if(recipeDao.selectRecipeById(rid).isEmpty()) throw new RecipeNotFound(rid);
+        if(profileDao.selectProfileById(uid).isEmpty()) throw new ProfileNotFound(uid);
         String uidString = uid.toString();
         String ridString = rid.toString();
         String sql = "DELETE FROM likes WHERE userId = ? AND recipeId = ?";
